[css-grid] Fix track sizing for auto minimum under min-content contraint
When a grid track has an auto min track sizing function and the
grid container is being sized under a min/max-content constraint,
the track’s base size should be set to the limited min/max-content
contribution.

This patch does not exactly follow the spec because Blink does not
perform two independent layout passes for min-content and max-content.
It lays out only once, and assigns the base size as the min-content
and the growth limit as the max-content. Therefore, this patch sets
the base size to the limited min-content contribution, even when
the grid container is being sized under a max-content constraint.
The result appears to be the same, as checked by the test.

Spec: https://drafts.csswg.org/css-grid/#algo-single-span-items

BUG=599042

TEST=external/wpt/css/css-grid/layout-algorithm/grid-sizing-under-min-max-content-constraints.html

Change-Id: I158bcdede2e878ab171964cdb9ddd25542ebd887